+++ core/src/main/kotlin/net/corda/core/serialization/CordaCustomSerializer.kt
@@ -1,0 +1,8 @@
+package net.corda.core.serialization
+
+@Target(AnnotationTarget.CLASS)
+annotation class CordaCustomSerializer
+
+@Target(AnnotationTarget.CLASS)
+annotation class CordaCustomSerializerProxy
+